front: refactor train header's service part into its own component
It should have been one from the beginning to lower the amount of code inside what has become a very long file anyway.
