+++ front/src/modules/trainHeader/ExpandedTrainForm.tsx
@@ -4,7 +4,6 @@ import {
   Button,
   ComboBox,
   DatePicker,
-  DurationInput,
   Input,
   Select,
   Switch,
@@ -12,11 +11,9 @@ import {
   useDefaultComboBox,
   type CalendarSlot,
 } from '@osrd-project/ui-core';
-import { ChevronUp } from '@osrd-project/ui-icons';
 import { isEqual } from 'lodash';
 import { useTranslation } from 'react-i18next';
 
-import type { PacedTrain } from 'applications/operationalStudies/types';
 import type {
   LightRollingStockWithLiveries,
   PathfindingResult,
@@ -32,17 +29,13 @@ import useCategoryOptions, {
 } from 'modules/rollingStock/hooks/useCategoryOptions';
 import useFilterRollingStock from 'modules/rollingStock/hooks/useFilterRollingStock';
 import type { Train } from 'reducers/osrdconf/types';
-import { Duration, MAX_DURATION_MS } from 'utils/duration';
+import { Duration } from 'utils/duration';
 import { usePrevious } from 'utils/hooks/state';
 import { kmhToMs } from 'utils/physics';
-import { findExceptionInPacedTrainByOccurrenceId } from 'utils/trainExceptions';
-import { isOccurrenceId } from 'utils/trainId';
 import { createFixedSelectOptions, createStandardSelectOptions } from 'utils/uiCoreHelpers';
 
-import ExtraOccurrenceForm from './ExtraOccurrenceForm';
-import ExtraOccurrenceRow from './ExtraOccurrenceRow';
-import { ServiceChangeWarningDialog } from './ServiceChangeWarningDialog';
 import type { ExtraOccurrencesChanges } from './TrainHeader';
+import TrainServiceForm, { computeServiceTimingError } from './TrainServiceForm';
 
 // TODO: Passing `undefined` to DatePicker's selectableSlot prop should mean this
 export const ANY_DATE_SLOT: CalendarSlot = { start: new Date(0), end: null };
@@ -58,7 +51,7 @@ export type ExpandedTrainFormProps = {
   onItineraryOpened: () => void;
 };
 
-type TrainFieldsState = {
+export type TrainFieldsState = {
   train_name: string;
   speed_limit_tag: string | null;
   constraint_distribution: ConstraintDistribution;
@@ -75,25 +68,17 @@ type TrainFieldsState = {
   service_changed_confirmed: boolean;
 };
 
-type initialSpeedProblem = 'INVALID_NUMBER' | 'ROUNDING' | 'TOO_HIGH' | null;
+export type InitialSpeedError = 'INVALID_NUMBER' | 'ROUNDING' | 'TOO_HIGH' | null;
 
 function computeInitialSpeedError(
   initialSpeed: string | null,
   rollingStock?: LightRollingStockWithLiveries
-): initialSpeedProblem {
+): InitialSpeedError {
   if (!initialSpeed) return 'INVALID_NUMBER';
   const floatInitialSpeed = Number.parseFloat(initialSpeed);
   if (!isFinite(floatInitialSpeed) || floatInitialSpeed < 0) return 'INVALID_NUMBER';
   if (Math.round(floatInitialSpeed * 10) / 10 !== floatInitialSpeed) return 'ROUNDING';
   if (rollingStock && kmhToMs(floatInitialSpeed) > rollingStock.max_speed) return 'TOO_HIGH';
-  return null;
-}
-
-type ServiceTimingError = 'TOO_LOW' | 'TOO_HIGH' | null;
-
-function computeServiceTimingError(serviceTiming?: number): ServiceTimingError {
-  if (!serviceTiming) return 'TOO_LOW';
-  if (serviceTiming > MAX_DURATION_MS) return 'TOO_HIGH';
   return null;
 }
 
@@ -227,18 +212,8 @@ const ExpandedTrainForm = ({
     resetSuggestions: resetRollingStockSuggestions,
   } = useDefaultComboBox(rollingStocks, getRollingStockLabel);
 
-  const pacedTrain = train.paced ? (train as PacedTrain) : null;
-  const occurrenceId = isOccurrenceId(train.id) ? train.id : null;
-  const exception =
-    occurrenceId && pacedTrain
-      ? findExceptionInPacedTrainByOccurrenceId(occurrenceId, pacedTrain)
-      : null;
-
   const fieldsFromTrain = useMemo(() => getFieldsFromTrain(train), [train]);
   const [fields, setFields] = useState<TrainFieldsState>(fieldsFromTrain);
-  const [extraOccurrencesVisible, setExtraOccurrencesVisible] = useState(false);
-  const extraOccurrences =
-    train.paced?.exceptions?.filter((exp) => exp.occurrence_index === undefined) ?? [];
 
   const selectedRollingStock = useMemo(
     () => rollingStocks.find((rs) => rs.name === fields.rolling_stock_name),
@@ -374,50 +349,6 @@ const ExpandedTrainForm = ({
     [fields.initial_speed, selectedRollingStock, computeInitialSpeedError]
   );
 
-  const isPacedTrain = !!pacedTrain;
-
-  const serviceCadenceError = useMemo(
-    () => (isPacedTrain ? computeServiceTimingError(fields.service_cadence) : null),
-    [isPacedTrain, fields.service_cadence]
-  );
-
-  const serviceWindowError = useMemo(
-    () => (isPacedTrain ? computeServiceTimingError(fields.service_window) : null),
-    [isPacedTrain, fields.service_window]
-  );
-
-  const erroneousFields = useMemo(
-    () =>
-      !fields.train_name ||
-      !fields.initial_speed ||
-      !!initialSpeedError ||
-      !!serviceCadenceError ||
-      !!serviceWindowError,
-    [
-      fields.train_name,
-      fields.initial_speed,
-      initialSpeedError,
-      serviceCadenceError,
-      serviceWindowError,
-    ]
-  );
-
-  const toggleBand = (
-    <div className="toggle-band">
-      <button className="header-toggle" onClick={() => onCollapse()} disabled={erroneousFields}>
-        <ChevronUp />
-      </button>
-    </div>
-  );
-
-  const isServiceChangeWarningDialogVisible = useMemo(
-    () =>
-      !fields.service_changed_confirmed &&
-      (fields.service_cadence?.valueOf() !== fieldsFromTrain.service_cadence?.valueOf() ||
-        fields.service_window?.valueOf() !== fieldsFromTrain.service_window?.valueOf()),
-    [train.paced, fields, fieldsFromTrain]
-  );
-
   const revertServiceChange = useCallback(() => {
     setFields({
       ...fields,
@@ -426,122 +357,21 @@ const ExpandedTrainForm = ({
     });
   }, [fields, fieldsFromTrain]);
 
-  const confirmServiceChange = useCallback(() => {
-    onFieldImmediateChange('service_changed_confirmed', true);
-  }, [onFieldImmediateChange]);
-
   return (
     <div className="train-header expanded-train-form">
-      {pacedTrain && (
-        <div className="train-service">
-          {toggleBand}
-          {occurrenceId ? (
-            <div className="train-occurrence">
-              <div className="train-paced-kind">
-                {t('manageTrainSchedule.trainHeader.serviceOccurrence')}
-                {exception && ' ≠'}
-              </div>
-            </div>
-          ) : (
-            <>
-              <div className="train-service-form">
-                <div className="train-paced-kind">
-                  {t('manageTrainSchedule.trainHeader.serviceModelTrain')}
-                </div>
-                <div className="train-service-cadence">
-                  <DurationInput
-                    id="train-header-service-cadence-input"
-                    small
-                    units={['m']}
-                    padChar="0"
-                    label={t('manageTrainSchedule.trainHeader.form.serviceCadence')}
-                    value={fields.service_cadence ?? 1_800_000} // 30m
-                    onChange={(ms) => onFieldImmediateChange('service_cadence', ms)}
-                    statusWithMessage={
-                      serviceCadenceError
-                        ? {
-                            status: 'error',
-                            message:
-                              serviceCadenceError === 'TOO_HIGH'
-                                ? t('manageTrainSchedule.errorMessages.intervalTooHigh')
-                                : t('manageTrainSchedule.errorMessages.intervalTooLow'),
-                          }
-                        : undefined
-                    }
-                  />
-                </div>
-                <div className="train-service-window">
-                  <DurationInput
-                    id="train-header-service-window-input"
-                    small
-                    units={['h', 'm']}
-                    padChar="0"
-                    label={t('manageTrainSchedule.trainHeader.form.serviceWindow')}
-                    value={fields.service_window ?? 2 * 3_600_000} // 2h00m
-                    onChange={(ms) => onFieldImmediateChange('service_window', ms)}
-                    statusWithMessage={
-                      serviceWindowError
-                        ? {
-                            status: 'error',
-                            message:
-                              serviceWindowError === 'TOO_HIGH'
-                                ? t('manageTrainSchedule.errorMessages.timeWindowTooHigh')
-                                : t('manageTrainSchedule.errorMessages.timeWindowTooLow'),
-                          }
-                        : undefined
-                    }
-                  />
-                </div>
-                <div className="actions">
-                  <Button
-                    label={t('manageTrainSchedule.trainHeader.form.extraOccurrences', {
-                      count: extraOccurrences.length,
-                    })}
-                    variant="Quiet"
-                    onClick={() => {
-                      setExtraOccurrencesVisible(!extraOccurrencesVisible);
-                    }}
-                    size="small"
-                  />
-                </div>
-              </div>
-
-              {extraOccurrencesVisible && (
-                <div className="train-extra-occurrences">
-                  <ExtraOccurrenceForm
-                    addedExceptionDate={fields.added_exception_date}
-                    setAddedExceptionDate={(newAddedExceptionDate) =>
-                      onFieldChange('added_exception_date', newAddedExceptionDate)
-                    }
-                    onCreateAddedException={() => {
-                      onPersistTrain(train, {
-                        addedExceptions: [{ startTime: fields.added_exception_date }],
-                      });
-                    }}
-                  />
-                  <div className="train-extra-occurrences-list">
-                    {extraOccurrences
-                      .filter((occurrence) => occurrence.start_time?.value)
-                      .map((occurrence) => (
-                        <ExtraOccurrenceRow
-                          key={`${occurrence.id}-${occurrence.key}`}
-                          startTime={new Date(occurrence.start_time!.value)}
-                          onDelete={() => {
-                            // TODO_EXCEPTION: remove this when the exception migration will be done
-                            if (occurrence.id !== null && occurrence.id !== undefined) {
-                              onPersistTrain(train, { deletedAddedExceptionId: occurrence.id });
-                            }
-                          }}
-                        />
-                      ))}
-                  </div>
-                </div>
-              )}
-            </>
-          )}
-        </div>
+      {train.paced && (
+        <TrainServiceForm
+          train={train}
+          fields={fields}
+          fieldsFromTrain={fieldsFromTrain}
+          initialSpeedError={initialSpeedError}
+          onCollapse={onCollapse}
+          onFieldChange={onFieldChange}
+          onFieldImmediateChange={onFieldImmediateChange}
+          onPersistTrain={onPersistTrain}
+          revertServiceChange={revertServiceChange}
+        />
       )}
-      {!pacedTrain && toggleBand}
       <div className="train-form">
         <div className="train-name">
           <Input
@@ -732,13 +562,6 @@ const ExpandedTrainForm = ({
         />
       )}
       {categoryWarning && <Banner message={categoryWarning} />}
-      {isServiceChangeWarningDialogVisible && (
-        <ServiceChangeWarningDialog
-          exceptionsCount={train.paced!.exceptions.length}
-          onCancel={revertServiceChange}
-          onConfirm={confirmServiceChange}
-        />
-      )}
     </div>
   );
 };
